backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
SQLAlchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.5.2
email-validator==2.1.0.post1
passlib==1.7.4
bcrypt==3.2.2
python-jose==3.3.0
cryptography==41.0.7